fix: updates based on review comments
