+++ tests/model_registry/image_validation/test_verify_rhoai_images.py
@@ -17,7 +17,6 @@ from pytest_testconfig import config as py_config
 from tests.model_registry.constants import MR_INSTANCE_NAME, MR_OPERATOR_NAME, MR_POSTGRES_DEPLOYMENT_NAME_STR
 from tests.model_registry.image_validation.utils import validate_images
 from utilities.constants import Labels
-from utilities.general import validate_image_format
 
 LOGGER = structlog.get_logger(name=__name__)
 pytestmark = [pytest.mark.downstream_only, pytest.mark.skip_must_gather]
@@ -54,26 +53,15 @@ class TestAIHubResourcesImages:
     @pytest.mark.tier1
     def test_verify_async_job_image(
         self: Self,
-        async_upload_image: str,
+        async_job_pod: Pod,
         related_images_refs: set[str],
     ):
         """
-        Given the async upload job image from the model-registry-operator-parameters ConfigMap
-        When validating the image reference
+        Given an async upload job deployed using the image from the operator ConfigMap
+        When validating the actual container image on the resulting pod
         Then verify it uses registry.redhat.io, sha256 digest, and is listed in CSV relatedImages
         """
-        LOGGER.info(f"Validating async upload job image: {async_upload_image}")
-        validation_errors = []
-
-        is_valid, error_msg = validate_image_format(image=async_upload_image)
-        if not is_valid:
-            validation_errors.append(f"Async upload job image validation failed: {error_msg}")
-
-        if async_upload_image not in related_images_refs:
-            validation_errors.append(f"Async upload job image {async_upload_image} is not in CSV relatedImages")
-
-        if validation_errors:
-            pytest.fail("\n".join(validation_errors))
+        validate_images(pods_to_validate=[async_job_pod], related_images_refs=related_images_refs)
 
 
 @pytest.mark.parametrize(
